requirements/base.txt
awesome-slugify==1.6.5
bcrypt==3.1.7
Django==3.0.1
djangorestframework==3.11.0
Pillow==6.2.1
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
